% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResNatSeed-package.R
\docType{package}
\name{ResNatSeed-package}
\alias{ResNatSeed}
\alias{ResNatSeed-package}
\title{ResNatSeed: blslllsldlasdasdasdasdas}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Marco Pittarello \email{marco.pittarello1987@gmail.com} (\href{https://orcid.org/0000-0001-6748-8790}{ORCID})

Authors:
\itemize{
  \item Davide Barberis \email{davide.barberis@unito.it} (\href{https://orcid.org/?????}{ORCID})
  \item Michele Lonati (\href{https://orcid.org/?????}{ORCID})
  \item Giampiero Lombardi (\href{https://orcid.org/?????}{ORCID})
}

}
\keyword{internal}
