% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResNatSeed.R
\name{ResNatSeed}
\alias{ResNatSeed}
\title{Computation of Mixture Suitability Index (MSI) and Mixture Relaiability Index (MRI) for a mixture/donor grassland composition
in a restoration site}
\usage{
ResNatSeed(trainingDB = NULL, composition, elevation, slope, aspect)
}
\arguments{
\item{trainingDB}{default is NULL. When NULL the training dataset is derived from 4081 vegetation surveys
carried out in Piedmont Region, Italy. A customized training dataset can be generated with \code{\link[ResNatSeed]{trainingDB}}
and then provided to "trainingDB" argument.}

\item{composition}{dataframe with the mixture/donor grassland composition. The dataframe
consists of two columns: 1) species code abbreviated in CEP names format and 2) abundance
of the species. When "trainingDB" is NULL, the species CEP names are retrievable from
\code{\link[ResNatSeed]{cep.piem}} data. When "trainingDB" is not NULL, i.e. the user does not use
default settings, the training dataset to be indicated in the "trainingDB" argument must be firstly
generated through \code{\link[ResNatSeed]{trainingDB}}. In this case, species CEP names that may
be indicated in the mixture/donor grassland composition in the "composition" argument are retrievable from \code{\link[ResNatSeed]{trainingDB}}}

\item{elevation}{Elevation of restoration site, expressed in meters above sea level (m a.s.l.)}

\item{slope}{Slope of restoration site, expressed in degrees (°)}

\item{aspect}{Aspect of restoration site, expressed in degrees from North (°N)}
}
\value{
A list with three outputs:
\itemize{
\item \strong{DESCRIPTIVES}: dataframe with the descriptive information related to species used in the mixture/donor grassland composition:
\describe{
\item{\emph{cep.names}}{Species name in CEP format}
\item{\emph{species}}{Full species name}
\item{\emph{cases.number}}{Number of cases related to each species}
\item{\emph{min.ele}}{Minimum elevation at which a species has been detected in the training dataset}
\item{\emph{max.ele}}{Maximum elevation at which a species has been detected in the training dataset}
\item{\emph{min.slope}}{Minimum slope at which a species has been detected in the training dataset}
\item{\emph{max.slope}}{Maximum slope at which a species has been detected in the training dataset}
\item{\emph{min.south}}{Minimum southness value at which a species has been detected in the training dataset}
\item{\emph{max.south}}{Maximum southness value at which a species has been detected in the training dataset}
}
\item \strong{SURVEYED AND PREDICTED ABUNDANCE}: dataframe with the following information:
\describe{
\item{\emph{cep.names}}{Species name in CEP format}
\item{\emph{species}}{Full species name}
\item{\emph{predicted.Abundance}}{Abundance of each species predicted by the best Generalized Additive Model (GAM) based on elevation, slope, and aspect of restoration site provided by the user}
\item{\emph{predicted.AbundanceMax}}{Maximum abundance achievable by a species under the best possible ecological conditions, identified by all possible combinations of altitude, slope and southness. This maximum abundance is predicted by the best GAM and the limits of the three topographic variables are indicated in the output of the descriptives }
\item{\emph{ratio}}{Ratio between the predicted abundance and the maximum abundance achievable. This ratio indicates how far (ratio = 0) or close (ratio = 1) a species is from its ecological optimum. }
\item{\emph{R2.adj}}{R square adjusted of the best GAM}
\item{\emph{RMSE}}{Root Mean Square Error of the best GAM}
\item{\emph{Mixture.abundance}}{Abundance of a species indicated in the mixture/donor grassland composition}
\item{\emph{Expected.abundance}}{Expected abundance (the most achievable) of a species in a restoration site with the topographical features provided by the user}
}
\item \strong{INDEXES}: \describe{
\item{\emph{MSI}}{Mixture Suitability Index (MSI). The MSI ranges from 0 (bad) to 1 (optimal). This index ...}
\item{\emph{MRI}}{Mixture Relaiability Index (MRI). The MRI ranges from 0 (bad) to 1 (optimal). This index ...}
}
}
}
\description{
The function allows to compute the Mixture Suitability Index (MSI) and Mixture Relaiability Index (MRI) for a mixture/donor grassland composition
in a restoration site based on topographic features, i.e. elevation, slope and aspect.
}
