% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cep.piem.R
\docType{data}
\name{cep.piem}
\alias{cep.piem}
\title{List of plant species for the seed mixture or donor grassland composition with the default training database}
\format{
A dataframe with 258 rows and 2 variables:
\describe{
\item{species}{Species names according to Aeschimann et al 2004}
\item{cep.names}{Abbreviation of species names according to the Cornell Ecology Programs (CEP), which uses eight-letter abbreviations for species. See \code{\link[vegan]{make.cepnames}} for more details}
}
}
\usage{
data(cep.piem)
}
\description{
A dataframe with the 258 plant species and their name abbreviation (CEP names) of the Piedmont Region (North Western Italy)
suitable for modeling when the \code{data} argument of \code{\link[ResNatSeed]{trainingDB}} function
is NULL. From this list is possible to create the mixture or donor grassland composition database.
}
\references{
\itemize{
\item Aeschimann, D., Lauber, K., Moser, M. D., & Theurillat, J. P. (2004). Flora alpina (Flora of the Alps). Zanichelli, Bologna.
}
}
\keyword{datasets}
