% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResNatSeed-package.R
\docType{package}
\name{ResNatSeed-package}
\alias{ResNatSeed}
\alias{ResNatSeed-package}
\title{ResNatSeed: A tool for the assessment of the REStoration potential of NATive SEEDs based on topographic factors}
\description{
ResNatSeed is an R package that compute the suitability of a certain seed mixture with a site where the seeds would be sown for restoration purposes in a degraded site.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MarcoPittarello/ResNatSeed}
}

}
\author{
\strong{Maintainer}: Marco Pittarello \email{marco.pittarello1987@gmail.com} (\href{https://orcid.org/0000-0001-6748-8790}{ORCID})

Authors:
\itemize{
  \item Davide Barberis \email{davide.barberis@unito.it}
}

}
\keyword{internal}
