% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResNatSeed.R
\name{ResNatSeed}
\alias{ResNatSeed}
\title{ResNatSeed}
\usage{
ResNatSeed(trainingDB = NULL, composition, elevation, slope, aspect)
}
\arguments{
\item{sourceDB}{default is NULL. When NULL ....}
}
\value{

}
\description{

}
\references{

}
