% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainingDB.R
\name{trainingDB}
\alias{trainingDB}
\title{Customized ResNatSeed training database}
\usage{
trainingDB(data, spe.freq, min.spe.abundance = NULL)
}
\arguments{
\item{data}{dataframe with vegetation and topographical variables with columns arranged as shown in @details and vignettes.}

\item{spe.freq}{threshold of minimum frequency of each species in the surveys. It is advisable to set values greater than or equal to 30 to allow appropriate statistical modeling.}

\item{min.spe.abundance}{threshold of minimum abundance (greater than) of each species in each survey. When NULL the parameter is set to 0.}
}
\value{
a list containing two dataframes:\cr
dataframe with the list of species suitable for modeling and their species codes (CEP names) to be used
for the seed mixture or donor grassland composition.
\itemize{
\item \strong{"cep.names"}: a two-column dataframe reporting the full species names and their abbreviations in the
CEP format (see \code{\link[vegan]{make.cepnames}} for more details). These species are those that need to be
indicated (in CEP names format) in the seed mixture or donor grassland composition
(i.e. in the \code{composition} argument of \code{\link[ResNatSeed]{RestInd}}).
\item \strong{"trainingDB.ResNatSeed"}: a seven-column dataframe reporting the following information: Survey ID, elevation, slope, southness, species full names, species CEP names, and species abundance. This dataframe must be used in the \code{trainingDB} argument of \code{\link[ResNatSeed]{RestInd}} function.
}
}
\description{
This function creates a training database based on vegetation and topographical variables database provided
by the user, which must be specified in the \emph{trainingDB} argument of
the function \code{\link[ResNatSeed]{RestInd}}.\cr

see vignettes for additional details and examples
}
\details{
The format of the dataframe required in the \code{data} argument must strictly follow this column order:\cr

SurveyID   |    Elevation   |    Slope   |    Aspect   |    Species1   |    Species2   |    Species3   |    Species…\cr
\itemize{
\item \strong{SurveyID} should contain an alphanumeric coding to uniquely identify each survey
\item \strong{Elevation} expressed in meters above sea level (m a.s.l.)
\item \strong{Slope} expressed in degrees (°)
\item \strong{Aspect} expressed in degrees from North (°N) as it will be converted to "southness" to avoid circular variable issues in the statistical models through the \code{\link[ResNatSeed]{RestInd}} function (Chang et al., 2004)
\item \strong{Species...} columns containing each plant species abundances. Abundances must be numbers bounded between 0 and 100, which can be either a species relative abundance or a species cover (\emph{sensu} Pittarello et al., 2016; Verdinelli et al.,2022 or see \code{\link[iPastoralist]{vegetation_abundance}} and \href{ https://raw.githubusercontent.com/MarcoPittarello/iPastoralist/main/image/Wrkflw_abundance_conversion.png }{here} \cr
}
}
\references{
\itemize{
\item Chang, C., P. Lee, M. Bai, and T. Lin. 2004. Predicting the geographical distribution of plant communities in complex terrain–a case study in Fushian Experimental Forest, north- eastern Taiwan. Ecography. 27:577–588. doi:10.1111/j.0906- 7590.2004.03852.x
\item Pittarello, M., Probo, M., Lonati, M., Lombardi, G., 2016. Restoration of sub-alpine shrub-encroached grasslands through pastoral practices: effects on vegetation structure and botanical composition. Appl Veg Sci 19, 381–390. https://doi.org/10.1111/avsc.12222
\item Verdinelli, M., Pittarello, M., Caria, M.C., Piga, G., Roggero, P.P., Marrosu, G.M., Arrizza, S., Fadda, M.L., Lombardi, G., Lonati, M., Nota, G., Sitzia, M., Bagella, S., 2022. Congruent responses of vascular plant and ant communities to pastoral land-use abandonment in mountain areas throughout different biogeographic regions. Ecol Process 11, 35. https://doi.org/10.1186/s13717-022-00379-9
\item Pittarello M. (2022). iPastoralist: Management, conversion, and analyses of vegetation data derived from phytosociological and point-quadrat method surveys. R package version 0.0.0.9000. https://github.com/MarcoPittarello/iPastoralist.git
}
}
