% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainingDB.R
\name{trainingDB}
\alias{trainingDB}
\title{Custom ResNatSeed source database and list of plant species that can be indicated in the mixture botanical composition}
\usage{
trainingDB(data, spe.freq, min.spe.abundance = NULL)
}
\arguments{
\item{data}{a dataframe where rows are surveys and columns are species. Such a dataframe must have the columns in the following order: Survey ID, Elevation, Slope, Aspect, and then columns with plant species abundance. See @details for additional and more specific information.}

\item{spe.freq}{minimum number of times a species must appear in order to be modelled. It is advisable to set values greater than/equal to 30 to allow appropriate modelling.}

\item{min.spe.abundance}{value of abundance beyond which species will be modelled. When NULL the parameter is set to 0.}
}
\value{
a list containing two dataframes:\cr
\itemize{
\item 1- \strong{"cep.names"}: a two-column dataframe reporting the full species names and their abbreviations the CEP format (see \code{link[vegan]{make.cepnames}} for more details). These species are those that can be indicated (as CEP names) in mixture botanical composition.
\item 2- \strong{"sourceDB.ResNatSeed"}: a seven-column dataframe reporting the following information: Survey ID, elevation, slope, southness, species full names, species CEP names, and species abundance. This dataframe must be used in the “sourceDB” argument of the \code{\link[ResNatSeed]{ResNatSeed}} function.
}
}
\description{
This function is to be used in case you do not want to use the source database set by default in the function \code{\link[ResNatSeed]{ResNatSeed}}, which includes vegetation surveys carried out in the Region Piedmont, Italy. \cr
\itemize{
\item 1 - the list of plant species that can be indicated in the mixture botanical composition with associated their species codes in CEP names format (see \code{\link[vegan]{make.cepnames}} for more details)
\item 2 - a source database in the format requested as input in the \code{\link[ResNatSeed]{ResNatSeed}} function.
}
}
\details{
The format of the dataframe required in the "data" argument must looks as follow (the headings do not necessarily have to be the same as those indicated, but it is essential to maintain that order):\cr

SurveyID   Elevation   Slope   Aspect   Species1   Species2   Species3   Species…\cr
\itemize{
\item \strong{SurveyID} should contain an alphanumeric coding
\item \strong{Elevation} can be expressed in meters or other similar unit of measurement
\item \strong{Slope} is recommended to be expressed in degrees (°)
\item \strong{Aspect} must be expressed in degrees from North (°N) as it will be converted to "southness" to avoid circular variable issues in the models run with \code{\link{ResNatSeed}} (Chang et al., 2004)
\item \strong{Species} columns contains plant species abundance, either expressed as percentage, Species Relative Cover (SRA), Species Percentage Cover (\%SC)(see more details regarding the last two terminologies in \code{\link[iPastoralist]{vegetation_abundance}} or \href{ https://raw.githubusercontent.com/MarcoPittarello/iPastoralist/main/image/Wrkflw_abundance_conversion.png }{here} \cr
}
}
\references{
\itemize{
\item Chang, C., P. Lee, M. Bai, and T. Lin. 2004. Predicting the geographical distribution of plant communities in complex terrain–a case study in Fushian Experimental Forest, north- eastern Taiwan. Ecography. 27:577–588. doi:10.1111/j.0906- 7590.2004.03852.x
}
}
